database refactor in progress
