+++ app/src/androidTest/java/ch/epfl/culturequest/database/DatabaseTest.java
@@ -3,15 +3,16 @@ package ch.epfl.culturequest.database;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.fail;
 
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.util.Calendar;
 import java.util.Date;
-import java.util.GregorianCalendar;
-import java.util.HashMap;
 import java.util.List;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeoutException;
 
 import ch.epfl.culturequest.social.Image;
 import ch.epfl.culturequest.social.Post;
@@ -21,76 +22,90 @@ public class DatabaseTest {
 
     @Before
     public void setUp() {
-        HashMap<String, Object> map = new HashMap<>();
+        // Set up the database to run on the local emulator of Firebase
+        Database.setEmulatorOn();
 
-        HashMap<String, Post> user1Posts = new HashMap<>();
-        map.put("user1", user1Posts);
+        // clear the database before starting the following tests
+        Database.clearDatabase();
 
-        HashMap<String, Post> user2Posts = new HashMap<>();
-        map.put("user2", user2Posts);
-
-        Database.init(new MockDatabase(map));
-    }
-
-    @Test
-    public void setAndGetWorks() {
-        Database.set("test", "test");
-        assertThat(Database.get("test").join(), is("test"));
-    }
-
-    @Test
-    public void setAndGetProfileWorks() {
-        Profile profile = new Profile("test", "test", "test", "test", "test", "test",null, 0);
-        Database.setProfile(profile);
-        assertThat(Database.getProfile("test").join(), is(profile));
+        // Initialize the database with some test profiles
     }
 
     @Test
     public void setAndGetImageWorks() {
         Image image = new Image("test", "test", "test", 0, "test");
-        Database.setImage(image);
-        assertThat(Database.getImage("test").join(), is(image));
+        try {
+            Database.setImage(image).get(5, java.util.concurrent.TimeUnit.SECONDS);
+            assertThat(Database.getImage("test").join(), is(image));
+        } catch (ExecutionException | InterruptedException | TimeoutException e) {
+            fail("Test failed because of an exception: " + e.getMessage());
+        }
+    }
+
+    @Test
+    public void setAndGetProfileWorks() {
+        Profile profile = new Profile("test", "test", "test", "test", "test", "test", List.of(), 0);
+        try {
+            Database.setProfile(profile).get(5, java.util.concurrent.TimeUnit.SECONDS);
+            assertThat(Database.getProfile("test").get(5, java.util.concurrent.TimeUnit.SECONDS), is(profile));
+        } catch (ExecutionException | InterruptedException | TimeoutException e) {
+            fail("Test failed because of an exception: " + e.getMessage());
+        }
     }
 
     @Test
     public void uploadPostWorks() {
         Post post = new Post("test", "user1", "test", "test", new Date(), 0, List.of());
-        boolean result = Database.uploadPost(post).join().get();
-        assertThat(result, is(true));
+        try {
+            assertThat(Database.uploadPost(post).get(5, java.util.concurrent.TimeUnit.SECONDS).get(), is(true));
+        } catch (ExecutionException | InterruptedException | TimeoutException e) {
+            fail("Test failed because of an exception: " + e.getMessage());
+        }
     }
 
     @Test
     public void removePostWorks() {
         Post post = new Post("test6", "user1", "test", "test", new Date(), 0, List.of());
-        boolean result = Database.uploadPost(post).join().get();
-        assertThat(result, is(true));
-        boolean result2 = Database.removePost(post).join().get();
-        assertThat(result2, is(true));
-        assertThat(Database.getPosts("user1", 10, 0).join().contains(post), is(false));
+        try {
+            assertThat(Database.uploadPost(post).get(5, java.util.concurrent.TimeUnit.SECONDS).get(), is(true));
+            assertThat(Database.removePost(post).get(5, java.util.concurrent.TimeUnit.SECONDS).get(), is(true));
+            assertThat(Database.getPosts("user1", 10, 0).get(5, java.util.concurrent.TimeUnit.SECONDS).contains(post), is(false));
+        } catch (ExecutionException | InterruptedException | TimeoutException e) {
+            fail("Test failed because of an exception: " + e.getMessage());
+        }
     }
 
     @Test
     public void getPostsWorks() {
-        Post post = new Post("test2", "user1", "test", "test", new Date(), 0, List.of());
+        Post post1 = new Post("test2", "user1", "test", "test", new Date(), 0, List.of());
         Post post2 = new Post("test3", "user1", "test", "test", new Date(), 0, List.of());
-        Database.uploadPost(post).join();
-        Database.uploadPost(post2).join();
-        assertThat(Database.getPosts("user1", 10, 0).join().get(0), is(post));
-        assertThat(Database.getPosts("user1", 10, 1).join().get(0), is(post2));
-        assertThat(Database.getPosts("user1", 1, 0).join().size(), is(1));
+        try {
+            Database.uploadPost(post1).get(5, java.util.concurrent.TimeUnit.SECONDS);
+            Database.uploadPost(post2).get(5, java.util.concurrent.TimeUnit.SECONDS);
+            Date date = new Date();
+            assertThat(Database.getPosts("user1", 10, 0).get(5, java.util.concurrent.TimeUnit.SECONDS).get(0), is(post1));
+            assertThat(Database.getPosts("user1", 10, 1).get(5, java.util.concurrent.TimeUnit.SECONDS).get(0), is(post2));
+            assertThat(Database.getPosts("user1", 1, 0).get(5, java.util.concurrent.TimeUnit.SECONDS).size(), is(1));
+        } catch (ExecutionException | InterruptedException | TimeoutException e) {
+            fail("Test failed because of an exception: " + e.getMessage());
+        }
     }
 
     @Test
     public void getPostsFeedWorks() {
-        Post post = new Post("test4", "user2", "test", "test", new Date(2023, 03, 28), 0, List.of());
+        Post post1 = new Post("test4", "user2", "test", "test", new Date(2023, 03, 28), 0, List.of());
         Post post2 = new Post("test5", "user2", "test", "test", new Date(2023, 03, 29), 0, List.of());
-        Database.uploadPost(post).join();
-        Database.uploadPost(post2).join();
-        assertThat(Database.getPostsFeed(List.of("user2"), 10, 0).join().get(0), is(post2));
-        assertThat(Database.getPostsFeed(List.of("user2")).join().get(0), is(post2));
-        assertThat(Database.getPostsFeed(List.of("user2"), 10).join().get(0), is(post2));
-        assertThat(Database.getPostsFeed(List.of("user2"), 10, 1).join().get(0), is(post));
-        assertThat(Database.getPostsFeed(List.of("user2"), 1, 0).join().size(), is(1));
+        try {
+            Database.uploadPost(post1).get(5, java.util.concurrent.TimeUnit.SECONDS);
+            Database.uploadPost(post2).get(5, java.util.concurrent.TimeUnit.SECONDS);
+            assertThat(Database.getPostsFeed(List.of("user2"), 10, 0).get(5, java.util.concurrent.TimeUnit.SECONDS).get(0), is(post2));
+            assertThat(Database.getPostsFeed(List.of("user2")).get(5, java.util.concurrent.TimeUnit.SECONDS).get(0), is(post2));
+            assertThat(Database.getPostsFeed(List.of("user2"), 10).get(5, java.util.concurrent.TimeUnit.SECONDS).get(0), is(post2));
+            assertThat(Database.getPostsFeed(List.of("user2"), 10, 1).get(5, java.util.concurrent.TimeUnit.SECONDS).get(0), is(post1));
+            assertThat(Database.getPostsFeed(List.of("user2"), 1, 0).get(5, java.util.concurrent.TimeUnit.SECONDS).size(), is(1));
+        } catch (ExecutionException | InterruptedException | TimeoutException e) {
+            fail("Test failed because of an exception: " + e.getMessage());
+        }
     }
 
     @Test
@@ -113,6 +128,12 @@ public class DatabaseTest {
         Database.removeLike(post, "user3").join();
         assertThat(Database.getPosts("user1", 10, 0).join().get(0).getLikes(), is(0));
         assertThat(Database.getPosts("user1", 10, 0).join().get(0).isLikedBy("user3"), is(false));
+    }
+
+    @After
+    public void tearDown() {
+        // clear the database after finishing the tests
+        Database.clearDatabase();
     }
 
 }
