+++ app/src/main/java/ch/epfl/culturequest/database/Database.java
@@ -1,8 +1,19 @@
 package ch.epfl.culturequest.database;
 
-import com.google.firebase.database.FirebaseDatabase;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.MutableData;
+import com.google.firebase.database.Transaction;
+
+import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
+import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
 
@@ -13,82 +24,161 @@ import ch.epfl.culturequest.social.Profile;
 
 
 /**
- * This class is used to implement the database
- * It is used to make the code more modular and to make it easier to change the database
+ * This class is the implementation of the database using Firebase
  */
 public class Database {
-    private static DatabaseInterface databaseInstance = new FireDatabase();
-    private static DatabaseInterface emulatorInstance = null;
-
-    /**
-     * This method is used to initialize the database instance as something else than the default (FireDatabase)
-     * This is useful for testing purposes
-     *
-     * @param database the database instance to be used
-     */
-    public static void init(DatabaseInterface database) {
-        Database.databaseInstance = database;
-    }
+    private static final FirebaseDatabase databaseInstance = FirebaseDatabase.getInstance();
+    private static boolean isEmulatorOn = false;
 
     public static void setEmulatorOn() {
-        if (emulatorInstance == null) {
-            FirebaseDatabase firebaseDatabase = FirebaseDatabase.getInstance();
-            firebaseDatabase.useEmulator("10.0.2.2", 9000);
-            databaseInstance = new FireDatabase(firebaseDatabase);
-            emulatorInstance = databaseInstance;
-        } else {
-            databaseInstance = emulatorInstance;
+        if (!isEmulatorOn) {
+            databaseInstance.useEmulator("10.0.2.2", 9000);
+            isEmulatorOn = true;
         }
     }
 
     public static void clearDatabase() {
-        databaseInstance.clearDatabase();
+        databaseInstance.getReference().setValue(null);
     }
 
-
-    public static CompletableFuture<AtomicBoolean> deleteProfile(String uid) {
-        return databaseInstance.deleteProfile(uid);
-    }
-
-    public static CompletableFuture<AtomicBoolean> set(String key, Object value) {
-        return databaseInstance.set(key, value);
-    }
-
-    public static CompletableFuture<Object> get(String key) {
-        return databaseInstance.get(key);
+    private static <T> CompletableFuture<T> getValue(DatabaseReference ref, Class<T> valueType) {
+        CompletableFuture<T> future = new CompletableFuture<>();
+        ref.get().addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                T value = task.getResult().getValue(valueType);
+                future.complete(value);
+            } else {
+                future.completeExceptionally(task.getException());
+            }
+        });
+        return future;
     }
 
     public static CompletableFuture<Profile> getProfile(String UId) {
-        return databaseInstance.getProfile(UId);
+        DatabaseReference usersRef = databaseInstance.getReference("users").child(UId);
+        return getValue(usersRef, Profile.class);
     }
 
-    public static CompletableFuture<AtomicBoolean> setProfile(Profile profile) {
-        return databaseInstance.setProfile(profile);
+    public static CompletableFuture<Image> getImage(String UId) {
+        DatabaseReference imagesRef = databaseInstance.getReference("images").child(UId);
+        return getValue(imagesRef, Image.class);
     }
 
     public static CompletableFuture<List<Profile>> getAllProfiles() {
-        return databaseInstance.getAllProfiles();
+        return getNumberOfProfiles().thenCompose(Database::getTopNProfiles);
     }
 
-
-    public static CompletableFuture<Image> getImage(String UId) {
-        return databaseInstance.getImage(UId);
+    public static CompletableFuture<AtomicBoolean> setImage(Image image) {
+        CompletableFuture<AtomicBoolean> future = new CompletableFuture<>();
+        databaseInstance.getReference("images").child(image.getUid()).setValue(image).addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                future.complete(new AtomicBoolean(true));
+            } else {
+                future.complete(new AtomicBoolean(false));
+            }
+        });
+        return future;
     }
 
-    public static CompletableFuture<AtomicBoolean> setImage(Image picture) {
-        return databaseInstance.setImage(picture);
+    public static CompletableFuture<AtomicBoolean> setProfile(Profile profile) {
+        CompletableFuture<AtomicBoolean> future = new CompletableFuture<>();
+        databaseInstance.getReference("users").child(profile.getUid()).setValue(profile).addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                future.complete(new AtomicBoolean(true));
+            } else {
+                future.complete(new AtomicBoolean(false));
+            }
+        });
+        return future;
     }
 
+    public static CompletableFuture<AtomicBoolean> deleteProfile(String uid) {
+        CompletableFuture<AtomicBoolean> future = new CompletableFuture<>();
+        DatabaseReference ref = databaseInstance.getReference("users").child(uid);
+        ref.removeValue((error, ref1) -> {
+            if (error == null) {
+                removeAllPosts(uid, future);
+            } else {
+                future.complete(new AtomicBoolean(false));
+            }
+        });
+        return future;
+    }
+
+    private static void removeAllPosts(String uid, CompletableFuture<AtomicBoolean> future) {
+        DatabaseReference ref = databaseInstance.getReference("posts").child(uid);
+        ref.removeValue((error, ref1) -> {
+            if (error == null) {
+                future.complete(new AtomicBoolean(true));
+            } else {
+                future.complete(new AtomicBoolean(false));
+            }
+        });
+    }
+
+    /**
+     * @param UId the user's id
+     * @return the rank of the user in the database with respect to their score
+     */
     public static CompletableFuture<Integer> getRank(String UId) {
-        return databaseInstance.getRank(UId);
+        DatabaseReference usersRef = databaseInstance.getReference("users");
+        CompletableFuture<Integer> future = new CompletableFuture<>();
+        getNumberOfProfiles().whenComplete((numberOfProfiles, e) -> {
+            usersRef.orderByChild("score").get().addOnCompleteListener(task -> {
+                if (task.isSuccessful()) {
+                    int rank = numberOfProfiles;
+                    for (DataSnapshot snapshot : task.getResult().getChildren()) {
+                        if (Objects.equals(snapshot.getKey(), UId)) {
+                            future.complete(rank);
+                            return;
+                        }
+                        rank--;
+                    }
+                    future.completeExceptionally(new RuntimeException("User not found"));
+                } else {
+                    future.completeExceptionally(task.getException());
+                }
+            });
+        });
+        return future;
     }
 
+    /**
+     * @return the number of profiles in the database
+     */
     public static CompletableFuture<Integer> getNumberOfProfiles() {
-        return databaseInstance.getNumberOfProfiles();
+        DatabaseReference usersRef = databaseInstance.getReference("users");
+        CompletableFuture<Integer> future = new CompletableFuture<>();
+        usersRef.get().addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                future.complete((int) task.getResult().getChildrenCount());
+            } else {
+                future.completeExceptionally(task.getException());
+            }
+        });
+        return future;
     }
 
+    /**
+     * @param n the number of profiles to get
+     * @return the top n profiles in the database with respect to their score
+     */
     public static CompletableFuture<List<Profile>> getTopNProfiles(int n) {
-        return databaseInstance.getTopNProfiles(n);
+        DatabaseReference usersRef = databaseInstance.getReference("users");
+        CompletableFuture<List<Profile>> future = new CompletableFuture<>();
+        usersRef.orderByChild("score").limitToLast(n).get().addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                List<Profile> profilesList = new ArrayList<>();
+                for (DataSnapshot snapshot : task.getResult().getChildren()) {
+                    Profile profile = snapshot.getValue(Profile.class);
+                    profilesList.add(profile);
+                }
+                future.complete(profilesList);
+            } else {
+                future.completeExceptionally(task.getException());
+            }
+        });
+        return future;
     }
 
     /**
@@ -98,7 +188,16 @@ public class Database {
      * @return a CompletableFuture that will be completed when the upload is done
      */
     public static CompletableFuture<AtomicBoolean> uploadPost(Post post) {
-        return databaseInstance.uploadPost(post);
+        CompletableFuture<AtomicBoolean> future = new CompletableFuture<>();
+        DatabaseReference usersRef = databaseInstance.getReference("posts").child(post.getUid()).child(String.valueOf(post.getPostid()));
+        usersRef.setValue(post).addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                future.complete(new AtomicBoolean(true));
+            } else {
+                future.complete(new AtomicBoolean(false));
+            }
+        });
+        return future;
     }
 
     /**
@@ -108,7 +207,16 @@ public class Database {
      * @return a CompletableFuture that will be completed when the removal is done
      */
     public static CompletableFuture<AtomicBoolean> removePost(Post post) {
-        return databaseInstance.removePost(post);
+        CompletableFuture<AtomicBoolean> future = new CompletableFuture<>();
+        DatabaseReference usersRef = databaseInstance.getReference("posts").child(post.getUid()).child(post.getPostid());
+        usersRef.removeValue((error, ref1) -> {
+            if (error == null) {
+                future.complete(new AtomicBoolean(true));
+            } else {
+                future.complete(new AtomicBoolean(false));
+            }
+        });
+        return future;
     }
 
     /**
@@ -120,7 +228,23 @@ public class Database {
      * @return a CompletableFuture that will be completed when the posts are retrieved
      */
     public static CompletableFuture<List<Post>> getPosts(String UId, int limit, int offset) {
-        return databaseInstance.getPosts(UId, limit, offset);
+        CompletableFuture<List<Post>> future = new CompletableFuture<>();
+        DatabaseReference usersRef = databaseInstance.getReference("posts").child(UId);
+        usersRef.orderByChild("date").limitToLast(limit + offset).get().addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                List<Post> posts = new ArrayList<>();
+                int i = 0;
+                for (DataSnapshot snapshot : task.getResult().getChildren()) {
+                    Post post = snapshot.getValue(Post.class);
+                    posts.add(post);
+                    if (++i == limit) break;
+                }
+                future.complete(posts);
+            } else {
+                future.complete(List.of());
+            }
+        });
+        return future;
     }
 
     /**
@@ -132,7 +256,31 @@ public class Database {
      * @return a CompletableFuture that will be completed when the posts are retrieved
      */
     public static CompletableFuture<List<Post>> getPostsFeed(List<String> UIds, int limit, int offset) {
-        return databaseInstance.getPostsFeed(UIds, limit, offset);
+        //List of posts to retrieve from each user
+        CompletableFuture<List<Post>>[] futures = UIds.stream().map((uid) -> getPosts(uid, limit, 0)).toArray(CompletableFuture[]::new);
+
+        //Future to return
+        CompletableFuture<List<Post>> result = new CompletableFuture<>();
+
+        //Wait for all futures to complete
+        CompletableFuture.allOf(futures).whenComplete((v, e) -> {
+
+            //Merge all posts into one list
+            List<Post> posts = new ArrayList<>();
+            for (CompletableFuture<List<Post>> future : futures) {
+                posts.addAll(future.join());
+            }
+
+            posts.sort(Comparator.comparing(Post::getDate).reversed());
+            posts = posts.subList(offset, Math.min(posts.size(), limit + offset));
+            result.complete(posts);
+
+        }).exceptionally(e -> {
+            result.complete(List.of());
+            return null;
+        });
+
+        return result;
     }
 
     /**
@@ -143,7 +291,7 @@ public class Database {
      * @return a CompletableFuture that will be completed when the posts are retrieved
      */
     public static CompletableFuture<List<Post>> getPostsFeed(List<String> UIds, int limit) {
-        return databaseInstance.getPostsFeed(UIds, limit);
+        return getPostsFeed(UIds, limit, 0);
     }
 
     /**
@@ -153,7 +301,7 @@ public class Database {
      * @return a CompletableFuture that will be completed when the posts are retrieved
      */
     public static CompletableFuture<List<Post>> getPostsFeed(List<String> UIds) {
-        return databaseInstance.getPostsFeed(UIds);
+        return getPostsFeed(UIds, 100, 0);
     }
 
     /**
@@ -164,25 +312,100 @@ public class Database {
      * @return a CompletableFuture that will be completed when the posts are retrieved
      */
     public static CompletableFuture<Post> addLike(Post post, String UId) {
-        return databaseInstance.addLike(post, UId);
+        return changeLike(post, UId, true);
     }
 
     /**
-     * This method is used to get the posts of a user's followings
-     *
-     * @param post the post to be liked
-     * @param uid  the user's id
-     * @return a CompletableFuture that will be completed when the posts are retrieved
+     * @param post the post to remove the like from
+     * @param UId  the id of the user who liked the post
+     * @return a future that will return true if the like was removed successfully, false otherwise
      */
-    public static CompletableFuture<Post> removeLike(Post post, String uid) {
-        return databaseInstance.removeLike(post, uid);
+    public static CompletableFuture<Post> removeLike(Post post, String UId) {
+        return changeLike(post, UId, false);
+    }
+
+    private static CompletableFuture<Post> changeLike(Post post, String UId, boolean add) {
+        CompletableFuture<Post> future = new CompletableFuture<>();
+        DatabaseReference usersRef = databaseInstance.getReference("posts").child(post.getUid()).child(post.getPostid());
+
+        usersRef.runTransaction(handler(future, UId, add));
+
+        return future;
+    }
+
+    private static Transaction.Handler handler(CompletableFuture<Post> future, String UId, boolean add) {
+        return new Transaction.Handler() {
+            @NonNull
+            @Override
+            public Transaction.Result doTransaction(@NonNull MutableData mutableData) {
+                Post dbPost = mutableData.getValue(Post.class);
+
+                if (dbPost == null) {
+                    future.completeExceptionally(new RuntimeException("Post not found"));
+                    return Transaction.abort();
+                }
+
+                if (add) {
+                    dbPost.addLike(UId);
+                } else {
+                    dbPost.removeLike(UId);
+                }
+
+                mutableData.setValue(dbPost);
+                return Transaction.success(mutableData);
+            }
+
+            @Override
+            public void onComplete(@Nullable DatabaseError databaseError, boolean b, @Nullable DataSnapshot dataSnapshot) {
+                if (databaseError != null || dataSnapshot == null) {
+                    future.complete(null);
+                } else {
+                    future.complete(dataSnapshot.getValue(Post.class));
+                }
+            }
+        };
     }
 
     public static CompletableFuture<Follows> addFollow(String follower, String followed) {
-        return databaseInstance.addFollow(follower, followed);
+        return changeFollow(follower, followed, true);
     }
 
     public static CompletableFuture<Follows> removeFollow(String follower, String followed) {
-        return databaseInstance.removeFollow(follower, followed);
+        return changeFollow(follower, followed, false);
     }
+
+    private static CompletableFuture<Follows> changeFollow(String follower, String followed, boolean follow) {
+        CompletableFuture<Follows> future = new CompletableFuture<>();
+        DatabaseReference followsRef = databaseInstance.getReference("follows").child(follower);
+
+        followsRef.runTransaction(new Transaction.Handler() {
+            @NonNull
+            @Override
+            public Transaction.Result doTransaction(@NonNull MutableData currentData) {
+                Follows follows = currentData.getValue(Follows.class);
+                if (follows == null) {
+                    follows = new Follows(List.of());
+                }
+                if (follow) {
+                    follows.addFollowed(followed);
+                } else {
+                    follows.removeFollowed(followed);
+                }
+                currentData.setValue(follows);
+                return Transaction.success(currentData);
+            }
+
+            @Override
+            public void onComplete(DatabaseError error, boolean committed, DataSnapshot currentData) {
+                if (error != null) {
+                    future.completeExceptionally(error.toException());
+                } else {
+                    future.complete(currentData.getValue(Follows.class));
+                }
+            }
+        });
+
+        return future;
+    }
+
 }
