+++ gradle.properties
@@ -25,3 +25,5 @@ org.gradle.daemon=true
 org.gradle.caching=true
 org.gradle.parallel=true
 org.gradle.configureondemand=true
+android.defaults.buildfeatures.buildconfig=true
+android.nonFinalResIds=false
